Change to Props + WebSockets
When having multiple components watch for messages on the same websocket they can block each other. Vue 3 uses Props, so updated the code to Props, resulting in cleaner format and more responsive actions.
