--- distvue/src/services/ws.js
@@ -1,2 +1,0 @@
-const WS = new WebSocket('ws://localhost:8081/ws');
-export default WS;
